Update type to standard C99 types

+++ kernel/libc/mem.c
@@ -1,12 +1,12 @@
 #include "mem.h"
 
 // this should be computed at linking, but hardcoding this is fine for now, the kernel starts at 0x1000
-u32 free_mem_address = 0x10000;
+uint32_t free_mem_address = 0x10000;
 
 // set memory at specified location
-void mem_set(u8* destination, u8 value, u32 length)
+void mem_set(uint8_t* destination, uint8_t value, uint32_t length)
 {
-    u8* temp = (u8*)destination;
+    uint8_t* temp = (uint8_t*)destination;
     
     while (length != 0)
     {
@@ -17,7 +17,7 @@ void mem_set(u8* destination, u8 value, u32 length)
 }
 
 // copy nbytes of value from 1 location to another location
-void mem_copy(u8* source, u8* destination, int nbytes)
+void mem_copy(uint8_t* source, uint8_t* destination, int nbytes)
 {
     for (int i = 0; i < nbytes; i++)
     {
@@ -26,7 +26,7 @@ void mem_copy(u8* source, u8* destination, int nbytes)
 }
 
 // can be used to request an aligned page, it will also return the real, physical address, for later use
-u32 kmalloc(u32 size, int align, u32* physical_address)
+uint32_t kmalloc(size_t size, int align, uint32_t* physical_address)
 {
     // pages are aligned to 4kb or 0x1000
     if (align == 1 && (free_mem_address & 0xFFFFF000))
@@ -38,7 +38,7 @@ u32 kmalloc(u32 size, int align, u32* physical_address)
     if (physical_address)
         *physical_address = free_mem_address;
 
-    u32 loc = free_mem_address; // location
+    uint32_t loc = free_mem_address; // location
     free_mem_address += size; // increment free address pointer to outside of specified size
 
     return loc;
