+++ kernel/libc/mem.h
@@ -1,10 +1,11 @@
 #ifndef MEM_H
 #define MEM_H
 
-#include "../cpu/types.h"
+#include <stdint.h>
+#include <stddef.h>
 
-void mem_set(u8* destination, u8 value, u32 length);
-void mem_copy(u8* source, u8* destination, int nbytes);
-u32 kmalloc(u32 size, int align, u32* physical_address); // currently there's no kfree() yet
+void mem_set(uint8_t* destination, uint8_t value, uint32_t length);
+void mem_copy(uint8_t* source, uint8_t* destination, int nbytes);
+uint32_t kmalloc(size_t size, int align, uint32_t* physical_address); // currently there's no kfree() yet
 
 #endif
